Move register data process functions to separate class.
